requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
Automat==22.10.0
black==23.10.1
buildtools==1.0.6
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
constantly==23.10.4
debugpy==1.8.0
decorator==5.1.1
docopt==0.6.2
executing==2.0.1
fastapi==0.104.1
furl==2.1.3
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.8
orderedmultidict==1.0.1
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
redo==2.0.4
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
tornado==6.3.3
traitlets==5.13.0
Twisted==23.10.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
wcwidth==0.2.9
zope.interface==6.1
